Change log handler to include file and line
Log handlers were previously required to acquire the file name and line
number from the file member in the parser. Providing both to the handler
is more intuitive. When multiple buffers become supported in a future
release it will also be more straightforward to the developer that the
file name is only correct while the log handler is executing.